upload semi-broken codebase requiring refactoring in packages
